spi_opencores: use RFEMPTY to determine end of transaction
The spi-simple SPI core has both a FIFO for buffering data and a
transfer register that's used for shifting bits in/out.  The WREMPTY
status bit only indicates that the write FIFO is empty, not that
everything has been shifted out because the last byte may be in the
transfer register.

In order to determine that a transfer has completed and a byte is
ready to read, it's better to test the RFEMPTY bit; this bit will
clear when there's a byte ready to be read.

Reported-by: Yann Vernier <yann.vernier@orsoc.se>
Signed-off-by: Jonas Bonn <jonas@southpole.se>