spi_opencores: wait for data ready, not send begun
WFEMPTY is about the write FIFO; when it's empty, that means the write
has been moved to the shift register itself. It does not mean the SPI
transfer has completed. Data shifted in is moved to a second fifo, with
with status flags RFEMPTY and RFFULL. This change waits before read to
check that there is data ready to read.